ENH: lower case module namesand Wsets renamed to set_operations
