+++ libpysal/weights/__init__.py
@@ -4,5 +4,5 @@ from .contiguity import *
 from .spintW import *
 from .util import *
 from .user import *
-from .Wsets import *
+from .set_operations import *
 from .spatial_lag import *
