+++ libpysal/weights/set_operations.py
