+++ libpysal/weights/tests/test_Wsets.py
@@ -1,17 +1,17 @@
-"""Unit test for Wsets module."""
+"""Unit test for set_operations module."""
 import unittest
 from ..util import lat2W, block_weights
-from .. import Wsets
+from .. import set_operations
 import numpy as np
 
-class TestWsets(unittest.TestCase):
-    """Unit test for Wsets module."""
+class Testset_operations(unittest.TestCase):
+    """Unit test for set_operations module."""
 
     def test_w_union(self):
         """Unit test"""
         w1 = lat2W(4, 4)
         w2 = lat2W(6, 4)
-        w3 = Wsets.w_union(w1, w2)
+        w3 = set_operations.w_union(w1, w2)
         self.assertEqual(w1[0], w3[0])
         self.assertEqual(set(w1.neighbors[15]), set([11, 14]))
         self.assertEqual(set(w2.neighbors[15]), set([11, 14, 19]))
@@ -21,7 +21,7 @@ class TestWsets(unittest.TestCase):
         """Unit test"""
         w1 = lat2W(4, 4)
         w2 = lat2W(6, 4)
-        w3 = Wsets.w_union(w1, w2)
+        w3 = set_operations.w_union(w1, w2)
         self.assertEqual(w1[0], w3[0])
         self.assertEqual(set(w1.neighbors[15]), set([11, 14]))
         self.assertEqual(set(w2.neighbors[15]), set([11, 14, 19]))
@@ -31,7 +31,7 @@ class TestWsets(unittest.TestCase):
         """Unit test"""
         w1 = lat2W(4, 4, rook=False)
         w2 = lat2W(4, 4, rook=True)
-        w3 = Wsets.w_difference(w1, w2, constrained=False)
+        w3 = set_operations.w_difference(w1, w2, constrained=False)
         self.assertNotEqual(w1[0], w3[0])
         self.assertEqual(set(w1.neighbors[15]), set([10, 11, 14]))
         self.assertEqual(set(w2.neighbors[15]), set([11, 14]))
@@ -41,7 +41,7 @@ class TestWsets(unittest.TestCase):
         """Unit test"""
         w1 = lat2W(4, 4, rook=False)
         w2 = lat2W(6, 4, rook=True)
-        w3 = Wsets.w_symmetric_difference(
+        w3 = set_operations.w_symmetric_difference(
             w1, w2, constrained=False)
         self.assertNotEqual(w1[0], w3[0])
         self.assertEqual(set(w1.neighbors[15]), set([10, 11, 14]))
@@ -52,7 +52,7 @@ class TestWsets(unittest.TestCase):
         """Unit test"""
         w1 = lat2W(6, 4)
         ids = list(range(16))
-        w2 = Wsets.w_subset(w1, ids)
+        w2 = set_operations.w_subset(w1, ids)
         self.assertEqual(w1[0], w2[0])
         self.assertEqual(set(w1.neighbors[15]), set([11, 14, 19]))
         self.assertEqual(set(w2.neighbors[15]), set([11, 14]))
@@ -63,7 +63,7 @@ class TestWsets(unittest.TestCase):
         w1.transform = 'R'
         w2 = block_weights(['r1', 'r2', 'r1', 'r1', 'r1', 'r2'])
         w2.transform = 'R'
-        wcs = Wsets.w_clip(w1, w2, outSP=True)
+        wcs = set_operations.w_clip(w1, w2, outSP=True)
         expected_wcs = np.array([[ 0.,  0.,0.33333333,0.33333333, 0.,0.],
                                  [ 0.,  0.,  0.,  0.,  0.,0. ],
                                  [ 0.2,  0.,  0.,  0.2,  0.2,0.],
@@ -72,11 +72,11 @@ class TestWsets(unittest.TestCase):
                                  [ 0.,  0.,  0.,  0.,  0.,0.]])
         np.testing.assert_array_equal(np.around(wcs.sparse.toarray(),decimals=8), expected_wcs)
 
-        wc = Wsets.w_clip(w1, w2, outSP=False)
+        wc = set_operations.w_clip(w1, w2, outSP=False)
         np.testing.assert_array_equal(wcs.sparse.toarray(), wc.full()[0])
 
 
-suite = unittest.TestLoader().loadTestsFromTestCase(TestWsets)
+suite = unittest.TestLoader().loadTestsFromTestCase(Testset_operations)
 
 if __name__ == '__main__':
     runner = unittest.TextTestRunner()
